YOLO_Segmentation/requirements.txt
librosa>=0.5.0
tensorflow==1.13.1
numpy>=1.12.1
pandas>=0.19.2
PyYAML>=3.11
SoundFile>=0.8.1
tqdm>=4.11.2
ipython==6.2.1
scipy>=1.3.0
dicttoxml>=1.7.4
wavio>=0.0.4
sounddevice=0.3.13
pascal_voc_writer==0.1.4
